third_party/WebKit/LayoutTests/fast/table/incorrect-colgroup-span-values.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../../resources/js-test.js"></script>
        <script type="text/javascript" charset="utf-8">
            if (window.testRunner)
                testRunner.dumpAsText();

            function test()
            {
                description("This test checks that invalid 'span' values are rejected.");

                // Remove "span" attribute (set it to null).
                document.getElementById("case1").removeAttribute("span");

                shouldBe("document.getElementById('case1').span", "1");
                shouldBe("document.getElementById('case2').span", "1");
                shouldBe("document.getElementById('case3').span", "1000");
                shouldBe("document.getElementById('case4').span", "1");
                shouldBe("document.getElementById('case5').span", "1");
                shouldBe("document.getElementById('case6').span", "1");
            }
        </script>
    </head>
    <body>
        <table>
            <colgroup id="case1" span="10000000000">
                <th></th>
            </colgroup>
        </table>
        <table>
            <colgroup id="case2" span="0">
                <th></th>
            </colgroup>
        </table>
        <table>
            <colgroup id="case3" span="10000000000">
                <th></th>
            </colgroup>
        </table>
        <table>
            <colgroup id="case4" span="-10000000000">
                <th></th>
            </colgroup>
        </table>
        <table>
            <colgroup id="case5" span="Szeged">
                <th></th>
            </colgroup>
        </table>
        <table>
            <colgroup>
                <col id="case6" span="...">
           </colgroup>
        </table>
        <script>
            test();
        </script>
    </body>
</html>
