third_party/blink/web_tests/fast/table/incorrect-colgroup-span-values-expected.txt

This test checks that invalid 'span' values are rejected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById('case1').span is 1
PASS document.getElementById('case2').span is 1
PASS document.getElementById('case3').span is 1000
PASS document.getElementById('case4').span is 1
PASS document.getElementById('case5').span is 1
PASS document.getElementById('case6').span is 1
PASS successfullyParsed is true

TEST COMPLETE
